Version 0.3
Version 0.3 - Game is in a near complete state.

- Map and movement have been reworked simplify adding design. Instead of a longer field, we now simply add a longer "strip" of tiles that will move with the player to give the illusion of movement while they're in the middle of the game, which is now only three screens wide. Movement has been adjusted that while within the "middle", we have an increasing counter called "D" (Delay) that allows us to give the QTEs exactly when we want. It also serves as the barrier to move on from the screen, as the player may only move to the "end" once they hit enough delay. Similarly, to go back to the start of the game, the player must decrease that delay. Both are critical for the illusion that they're running down a hallway.

- Ending has been implemented, for the "true" ending at least. Once the player reaches the end, movement will be paused and the device will "explode". After a few moments, the screen will cut to a scene of a planet in space next to the sun with randomly placed stars in the background, and a small part of the planet will blow up, while giving the game-over text. We use the explosion variable (used to increase the size of the explosion visual) as a delay for the text and a small part of the explosion on the planet.

We'll need to clean up the smaller visual bugs and look of the game still, and a "punishment" for failing the QTE should be implemented. We *could* simply have it auto fail, but it would be odd for the explosion to happen early. Conversely, forcing the player to wait the 20 seconds is tough. We could simply allow room for *a* mistake, and simply do a "trip" animation. We also need to adjust the QTE's such that they don't appear if one is still on screen, and instead delay.