Version 0.4
Version 0.4 - Bug Fixes have been made, and fine-tuning of the "fail" systems have been implemented. As it stands, messing up the QTE will cause you to suffer a penalty of 1 second for the failed one, and auto fail the rest (which would also give additional punishment.) We still need to balance it appropriately, but this should do the job fine. As it stands, the player can fail effectively "3" buttons before victory is more or less impossible.

I'd love to spend more time adjusting the QTE mechanics, I feel that I can make it less clunky, though I still need to adjust the sprites and animations appropriately.