Version 0.5
Version 0.5 

Title Screen has been added and implemented, as well as a cute transition from it to the game. Minor bug fixes have also been added as noticed. Various wild loops have been compartmentalized to make code reading easier and ensure ease of understanding what does what, when and why. The next update should hopefully be the last, and it'll cover a few more touch-ups to make the game look pretty, and add the player's actual sprite+tripping. 

Wish me luck! I say. Knowing that if you're reading this, it's already over and there's little point in reading something now that the event is over. I hope that I did okay, and that it was at least liked!