Version 1.0
First Release of the Game!

The game is in a fully playable state. Visuals have been overhauled, and various systems have been worked to make the gameplay feel as smooth as possible. Code has been reworked a few times to try and make it as reusable as possible. Ideally we'll be finished with the project with this release!